ENH: allow RegularSurface() store np.float32 values
As default, the numpy (masked) array is stored as 64 bit numpy,
in line with Roxar API storage. Here it is allowed to cast to 32
bit in import. The methods surface_from_grid() and surface_from_cube()
Still lack dtype, but it is easy to set with the property
srf.dtype = ...

+++ src/xtgeo/surface/_regsurf_export.py
@@ -101,7 +101,7 @@ def _export_irap_ascii(self, mfile):
         self._xinc,
         self._yflip * self._yinc,
         self._rotation,
-        vals,
+        vals.astype(np.float64),
         0,
     )
     if ier != 0:
@@ -199,7 +199,7 @@ def _export_irap_binary_cxtgeo(self, mfile):
         self._xinc,
         self._yflip * self._yinc,
         self._rotation,
-        vals,
+        vals.astype(np.float64),
         0,
     )
 
@@ -225,7 +225,7 @@ def export_ijxyz_ascii(self, mfile):
         self._yflip,
         self._ilines,
         self._xlines,
-        vals,
+        vals.astype(np.float64),
         0,
     )
 
@@ -324,7 +324,7 @@ def _export_zmap_ascii(self, mfile):
         scopy._yori,
         scopy._xinc,
         yinc,
-        vals,
+        vals.astype(np.float64),
         zmin,
         zmax,
         0,
@@ -361,7 +361,9 @@ def export_storm_binary(self, mfile):
         scopy._yori,
         scopy._xinc,
         yinc,
-        scopy.get_values1d(order="F", asmasked=False, fill_value=self.undef),
+        scopy.get_values1d(order="F", asmasked=False, fill_value=self.undef).astype(
+            np.float64
+        ),
         zmin,
         zmax,
         0,
@@ -416,7 +418,7 @@ def export_petromod_binary(self, mfile, pmd_dataunits):
     _cxtgeo.surf_export_petromod_bin(
         mfile.get_cfhandle(),
         dsc,
-        values,
+        values.astype(np.float64),
     )
 
     mfile.cfclose()
